Reconstruct THillFortWindow::Recalculate 0x4e7eb0 (79.70%, documented)
The 1613-byte body behind DoModal and the handler's redraw: it rebuilds
all seven upgrade rows from the current hero's army, prices each upgrade
through get_upgrade_cost scaled by the .rdata level table, accumulates
totalCost, drives every widget through BroadcastMessage, and settles both
the per-slot and the upgrade-all tri-states.

Claims this TU's three .rdata objects while it is here: the two icon
tables the states index (aphlf1y/1g/1r and aphlf4y/4g/4r) and the seven
float level multipliers, whose extent is bounded exactly - 0x63eb4c +
7*4 lands on the vftable at 0x63eb68.

Byte findings:
* The per-slot tri-state and UpgradeAllButtonState share one domain and
  one meaning: 0 = nothing to upgrade, 1 = affordable, 2 = too expensive,
  read straight out as an index into the two icon tables.
* The empty-slot arm dims the row's upgrade button with WIDGET_DIMMED
  when the caller asks for dimmed buttons and WIDGET_DIMMED_NODRAW when
  it does not - the `neg al` / `sbb eax,eax` / `and 0xfffff008` /
  `add 0x1000` idiom is that ternary.

Residual is one strength-reduction choice cascading through the slot
loop (retail homes the widget id and bases the slot pointer on
slot[i].count; this compile homes the index and bases it on cost[6]);
six spellings measured and named in the in-source note.

1061/1435 exact (73.9%), 60.12% fuzzy, executable matched 15.94%.

+++ include/hillfortwindow.h
@@ -123,7 +123,7 @@ public:
         char szCount[10];         // +0x00
         char szGoldCost[10];      // +0x0a
         char szResourceCost[12];  // +0x14
-        long cost[7];             // +0x20
+        int cost[7];              // +0x20
         int resourceIndex;        // +0x3c
         int type;                 // +0x40 (TCreatureType domain)
         int count;                // +0x44
@@ -145,7 +145,7 @@ public:
     void DoModal();
 
     TUpgradeSlot slot[armyGroup::ARMY_GROUP_SLOT_COUNT];   // +0x4c
-    long totalCost[armyGroup::ARMY_GROUP_SLOT_COUNT];      // +0x27c
+    int totalCost[armyGroup::ARMY_GROUP_SLOT_COUNT];       // +0x27c
     int UpgradeAllButtonState;                             // +0x298
     widget* RolloverWidget;                                // +0x29c
